+++ src/effects/mod.rs
@@ -5,8 +5,9 @@ mod greyscale;
 pub use blur::Blur;
 pub use dither::{
     ChannelMode, CmykScreenPreset, Color, Colour, ColourChannel, ColourHalftone,
-    ColourHalftoneMode, DiffusionAlgorithm, DiffusionKernel, DiffusionScan, DiffusionTap, DotShape,
-    ErrorDiffusion, Halftone, HalftoneChannel, HalftoneShape, NoiseAlgorithm, NoiseDither,
-    OrderedDither, Palette, Threshold, ThresholdMap, ThresholdRotation,
+    ColourHalftoneMode, ColourSpace, DiffusionAlgorithm, DiffusionErrorMode, DiffusionKernel,
+    DiffusionScan, DiffusionTap, DotShape, ErrorDiffusion, Halftone, HalftoneChannel,
+    HalftoneShape, NoiseAlgorithm, NoiseDither, OrderedDither, Palette, PaletteMatchMode,
+    Threshold, ThresholdMap, ThresholdRotation,
 };
 pub use greyscale::Greyscale;